refactor(demand-mgmt): move score_demand_item into a scoped tool pack
The executeTool inline-case ratchet (tool-registry.test) requires new MCP
tools to register in a scoped pack, not the mega-switch. Extract
score_demand_item into demand-scoring-pack.ts (definition + handler + grant),
register it in composeToolPacks, mirror the grant in tak/agent-grants.ts
(score_demand_item -> backlog_write), and add the pack's grant-drift +
presence test block. Behaviour unchanged.
