+++ apps/web/lib/mcp-tools.ts
@@ -12,7 +12,7 @@ import * as crypto from "crypto";
 import { lazyFs, lazyFsPromises, lazyPath, lazyChildProcess, lazyUtil, getCwd } from "@/lib/shared/lazy-node";
 import { slugify } from "@/lib/shared/slugify";
 import { mergeHappyPathStateIntoPlan, generateBuildId } from "@/lib/feature-build-types";
-import { BACKLOG_SOURCE_VALUES, BACKLOG_STATUS_VALUES, BACKLOG_WORK_TYPE_VALUES, DEMAND_SCORE_FRAMEWORKS, DEMAND_STAGE_VALUES, EPIC_STATUSES } from "@/lib/explore/backlog";
+import { BACKLOG_SOURCE_VALUES, BACKLOG_STATUS_VALUES, BACKLOG_WORK_TYPE_VALUES, EPIC_STATUSES } from "@/lib/explore/backlog";
 import {
   analyzePublicWebsiteBranding,
   fetchPublicWebsiteEvidence,
@@ -64,6 +64,7 @@ import { queueAwarenessPack } from "@/lib/mcp/packs/queue-awareness-pack";
 import { documentPack } from "@/lib/mcp/packs/document-pack";
 import { screenPack } from "@/lib/mcp/packs/screen-pack";
 import { nonprodLeasePack } from "@/lib/mcp/packs/nonprod-lease-pack";
+import { demandScoringPack } from "@/lib/mcp/packs/demand-scoring-pack";
 import { composeToolPacks } from "@/lib/mcp/tool-registry";
 import {
   createLicenseReadinessIssue,
@@ -450,7 +451,7 @@ function stringArray(value: unknown): string[] {
 // ─── Tool Registry ───────────────────────────────────────────────────────────
 // Scoped tool packs compose into the registry; mcp-tools.ts is the thin layer
 // over them (definitions spread into PLATFORM_TOOLS below; dispatch in executeTool).
-const TOOL_PACK_REGISTRY = composeToolPacks([deliberationSiemPack, runtimeCoordinationPack, workCapsulesPack, workbooksPack, feedbackPack, orgDecisionPack, professionDecisionPack, optimizationPack, marketingPack, workCapturePack, activityRoutingPack, selfUpgradePack, coworkerServiceCatalogPack, coworkerToolGrantPack, coworkerEstablishPack, coworkerMemoryPack, effortContextPack, coworkerGoalPack, subagentFanoutPack, mdmStewardshipPack, crmContactsPack, queueAwarenessPack, documentPack, screenPack, nonprodLeasePack]);
+const TOOL_PACK_REGISTRY = composeToolPacks([deliberationSiemPack, runtimeCoordinationPack, workCapsulesPack, workbooksPack, feedbackPack, orgDecisionPack, professionDecisionPack, optimizationPack, marketingPack, workCapturePack, activityRoutingPack, selfUpgradePack, coworkerServiceCatalogPack, coworkerToolGrantPack, coworkerEstablishPack, coworkerMemoryPack, effortContextPack, coworkerGoalPack, subagentFanoutPack, mdmStewardshipPack, crmContactsPack, queueAwarenessPack, documentPack, screenPack, nonprodLeasePack, demandScoringPack]);
 
 export const PLATFORM_TOOLS: ToolDefinition[] = [
   ...TOOL_PACK_REGISTRY.definitions,
@@ -536,29 +537,6 @@ export const PLATFORM_TOOLS: ToolDefinition[] = [
         size: { type: "string", enum: ["small", "medium", "large", "xlarge"], description: "T-shirt size estimate" },
       },
       required: ["itemId", "size"],
-    },
-    requiredCapability: "manage_backlog",
-    sideEffect: true,
-  },
-  {
-    name: "score_demand_item",
-    description:
-      "Set demand-scoring inputs on a backlog item and (re)compute its demandScore under a pluggable framework. Store the INPUTS; the score is derived. reach falls back to occurrenceCount and jobSize to the t-shirt effortSize when the explicit field is omitted. When the minimum inputs for the framework are present the item advances demandStage from raw to screened. Framework defaults to rice.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        itemId: { type: "string", description: "The item ID to score" },
-        framework: { type: "string", enum: [...DEMAND_SCORE_FRAMEWORKS], description: "Scoring framework. Defaults to rice (the seeded default). rice=(reach*impact*confidence)/jobSize; wsjf=(businessValue+timeCriticality+riskOpportunity)/jobSize; value_effort=value/jobSize; weighted=weighted sum." },
-        reach: { type: "integer", description: "RICE reach — users/instances affected per period. Falls back to occurrenceCount." },
-        impact: { type: "number", description: "RICE impact / value magnitude (e.g. 3=massive,2=high,1=medium,0.5=low,0.25=minimal)." },
-        confidence: { type: "number", description: "RICE confidence as a fraction (1.0 high, 0.8 medium, 0.5 low)." },
-        businessValue: { type: "integer", description: "WSJF Cost-of-Delay term — relative business value." },
-        timeCriticality: { type: "integer", description: "WSJF Cost-of-Delay term — relative time criticality." },
-        riskOpportunity: { type: "integer", description: "WSJF Cost-of-Delay term — relative risk reduction / opportunity enablement." },
-        jobSize: { type: "number", description: "Effort denominator (relative points). Falls back to the effortSize t-shirt map (small=1,medium=3,large=8,xlarge=20)." },
-        demandStage: { type: "string", enum: [...DEMAND_STAGE_VALUES], description: "Optional explicit funnel stage override (raw|screened|shaped|ready). Normally derived." },
-      },
-      required: ["itemId"],
     },
     requiredCapability: "manage_backlog",
     sideEffect: true,
@@ -5186,75 +5164,6 @@ export async function executeTool(
         success: true,
         entityId: itemId,
         message: `Sized ${itemId} as ${size}`,
-      };
-    }
-
-    case "score_demand_item": {
-      const itemId = String(params["itemId"] ?? "");
-      const item = await prisma.backlogItem.findUnique({ where: { itemId } });
-      if (!item) {
-        return { success: false, error: "Item not found", message: `Item ${itemId} not found` };
-      }
-      const { computeDemandScore } = await import("@/lib/demand/scoring");
-      const numOrKeep = (key: string, current: number | null): number | null =>
-        typeof params[key] === "number" ? (params[key] as number) : current;
-      const framework = ((): (typeof DEMAND_SCORE_FRAMEWORKS)[number] => {
-        const f = String(params["framework"] ?? "rice");
-        return (DEMAND_SCORE_FRAMEWORKS as readonly string[]).includes(f)
-          ? (f as (typeof DEMAND_SCORE_FRAMEWORKS)[number])
-          : "rice";
-      })();
-      // Merge supplied inputs over what's already stored (partial updates).
-      const inputs = {
-        reach: numOrKeep("reach", item.reach),
-        impact: numOrKeep("impact", item.impact),
-        confidence: numOrKeep("confidence", item.confidence),
-        businessValue: numOrKeep("businessValue", item.businessValue),
-        timeCriticality: numOrKeep("timeCriticality", item.timeCriticality),
-        riskOpportunity: numOrKeep("riskOpportunity", item.riskOpportunity),
-        jobSize: numOrKeep("jobSize", item.jobSize),
-        occurrenceCount: item.occurrenceCount,
-        effortSize: item.effortSize,
-      };
-      const result = computeDemandScore(inputs, framework);
-      // Advance the funnel: a scored item is at least `screened`. Respect an
-      // explicit stage override and never regress a manually-advanced stage.
-      const explicitStage = (DEMAND_STAGE_VALUES as readonly string[]).includes(String(params["demandStage"]))
-        ? (params["demandStage"] as (typeof DEMAND_STAGE_VALUES)[number])
-        : null;
-      const stageOrder = DEMAND_STAGE_VALUES as readonly string[];
-      const derivedStage = result.score !== null ? "screened" : (item.demandStage ?? "raw");
-      const currentIdx = item.demandStage ? stageOrder.indexOf(item.demandStage) : -1;
-      const derivedIdx = stageOrder.indexOf(derivedStage);
-      const nextStage = explicitStage ?? (derivedIdx > currentIdx ? derivedStage : item.demandStage ?? derivedStage);
-      await prisma.backlogItem.update({
-        where: { itemId },
-        data: {
-          reach: inputs.reach,
-          impact: inputs.impact,
-          confidence: inputs.confidence,
-          businessValue: inputs.businessValue,
-          timeCriticality: inputs.timeCriticality,
-          riskOpportunity: inputs.riskOpportunity,
-          jobSize: inputs.jobSize,
-          demandScore: result.score,
-          demandScoreFramework: framework,
-          demandScoreComputedAt: new Date(),
-          demandStage: nextStage,
-        },
-      });
-      return {
-        success: true,
-        entityId: itemId,
-        message:
-          result.score !== null
-            ? `Scored ${itemId}: ${framework} = ${result.score} (stage ${nextStage})`
-            : `Recorded inputs for ${itemId}; ${framework} not computable — missing ${result.missing.join(", ")}`,
-        demandScore: result.score,
-        framework,
-        demandStage: nextStage,
-        contributions: result.contributions,
-        missing: result.missing,
       };
     }
